hi_medei/samples/python/requirements.txt
click>=8.0.0
PyPDF2>=3.0.0
langchain-google-genai>=0.0.5
langgraph>=0.0.10
python-dotenv>=0.19.0
httpx>=0.24.0
pydantic>=2.0.0
sse-starlette>=1.6.5
starlette>=0.27.0
uvicorn>=0.24.0
PyJWT>=2.8.0
jwcrypto>=1.5.0
cryptography>=41.0.0
python-multipart>=0.0.6
typing-extensions>=4.7.0
aiofiles>=23.2.1
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.24.0 